Refactor state index
